ui: faculty placed in the place of departments in admin dashboard.
