--- react-frontend/src/features/departments/services/index.js
@@ -1,29 +1,0 @@
-import axios from "axios";
-
-const API_URL = "http://localhost:8888/api/departments";
-
-export const fetchDepartments = async () => {
-  try {
-    const response = await axios.get(API_URL);
-    return response;
-  } catch (e) {
-    console.error("while fecthing departments", e);
-  }
-};
-
-export const createDepartment = async (department) => {
-  try {
-    let res = await axios.post(API_URL, department);
-    return res;
-  } catch (e) {
-    console.error("while creating department", e);
-  }
-};
-
-export const fetchDepartmentFaculty = async (did) => {
-  return await axios.get(`${API_URL}/${did}/faculty`);
-}
-
-export const createDepartmentFaculty = async (did, data) => {
-  return await axios.post(`${API_URL}/${did}/faculty`, data);
-}
